mockenv/consumer/requirements.txt
confluent-kafka
psycopg2-binary
redis
